+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `sdwan_transport_wan_vpn_interface_cellular_feature_associate_tracker_feature` resource and data source
 - Add `sdwan_transport_wan_vpn_interface_gre_feature_associate_tracker_feature` resource and data source
 - Add `sdwan_transport_wan_vpn_interface_ipsec_feature_associate_tracker_feature` resource and data source
+- Add `sdwan_service_lan_vpn_interface_ethernet_feature_associate_tracker_feature` resource and data source
 
 ## 0.5.2