+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.4 (unreleased)
+
+- Delete configuration group when an error is encountered during creation, [link](https://github.com/CiscoDevNet/terraform-provider-sdwan/issues/390)
+
 ## 0.5.3
 
 - Fix template rendering issue with `sdwan_cisco_vpn_interface_cellular_feature_template` resource, [link](https://github.com/CiscoDevNet/terraform-provider-sdwan/issues/378)